ui: Move update_timeline_reaction to TimelineInnerState
