+++ crates/matrix-sdk-ui/src/timeline/inner/mod.rs
@@ -65,8 +65,8 @@ use super::{
     rfind_event_by_id, rfind_event_item,
     traits::RoomDataProvider,
     AnnotationKey, EventSendState, EventTimelineItem, InReplyToDetails, Message, Profile,
-    ReactionSenderData, RelativePosition, RepliedToEvent, TimelineDetails, TimelineItem,
-    TimelineItemContent, TimelineItemKind,
+    RelativePosition, RepliedToEvent, TimelineDetails, TimelineItem, TimelineItemContent,
+    TimelineItemKind,
 };
 
 mod state;
@@ -643,7 +643,7 @@ impl<P: RoomDataProvider> TimelineInner<P> {
                 // We're done, so also update the timeline
                 state.in_flight_reaction.remove(&annotation_key);
                 state.reaction_state.remove(&annotation_key);
-                update_timeline_reaction(&mut state, user_id, annotation, result)?;
+                state.update_timeline_reaction(user_id, annotation, result)?;
 
                 ReactionAction::None
             }
@@ -1138,99 +1138,4 @@ async fn fetch_replied_to_event(
         Err(e) => TimelineDetails::Error(Arc::new(e)),
     };
     Ok(res)
-}
-
-fn update_timeline_reaction(
-    state: &mut TimelineInnerState,
-    own_user_id: &UserId,
-    annotation: &Annotation,
-    result: &ReactionToggleResult,
-) -> Result<(), super::Error> {
-    if matches!(result, ReactionToggleResult::RedactSuccess) {
-        // We did a successful redaction, so no need to update the item
-        // because the reaction is already gone.
-        return Ok(());
-    }
-
-    let (remote_echo_to_add, local_echo_to_remove) = match result {
-        ReactionToggleResult::AddSuccess { event_id, txn_id } => (Some(event_id), Some(txn_id)),
-        ReactionToggleResult::AddFailure { txn_id } => (None, Some(txn_id)),
-        ReactionToggleResult::RedactSuccess => (None, None),
-        ReactionToggleResult::RedactFailure { event_id } => (Some(event_id), None),
-    };
-
-    let related = rfind_event_item(&state.items, |it| {
-        it.event_id().is_some_and(|it| it == annotation.event_id)
-    });
-
-    let Some((idx, related)) = related else {
-        // Event isn't found at all.
-        warn!("Timeline item not found, can't update reaction ID");
-        return Err(super::Error::FailedToToggleReaction);
-    };
-    let Some(remote_related) = related.as_remote() else {
-        error!("inconsistent state: reaction received on a non-remote event item");
-        return Err(super::Error::FailedToToggleReaction);
-    };
-    // Note: remote event is not synced yet, so we're adding an item
-    // with the local timestamp.
-    let reaction_sender_data = ReactionSenderData {
-        sender_id: own_user_id.to_owned(),
-        timestamp: MilliSecondsSinceUnixEpoch::now(),
-    };
-
-    let new_reactions = {
-        let mut reactions = remote_related.reactions.clone();
-        let reaction_group = reactions.entry(annotation.key.clone()).or_default();
-
-        // Remove the local echo from the related event.
-        if let Some(txn_id) = local_echo_to_remove {
-            let id = EventItemIdentifier::TransactionId(txn_id.clone());
-            if reaction_group.0.remove(&id).is_none() {
-                warn!(
-                    "Tried to remove reaction by transaction ID, but didn't \
-                     find matching reaction in the related event's reactions"
-                );
-            }
-        }
-        // Add the remote echo to the related event
-        if let Some(event_id) = remote_echo_to_add {
-            reaction_group.0.insert(
-                EventItemIdentifier::EventId(event_id.clone()),
-                reaction_sender_data.clone(),
-            );
-        };
-        if reaction_group.0.is_empty() {
-            reactions.remove(&annotation.key);
-        }
-
-        reactions
-    };
-    let new_related = related.with_kind(remote_related.with_reactions(new_reactions));
-
-    // Update the reactions stored in the timeline state
-    {
-        // Remove the local echo from reaction_map
-        // (should the local echo already be up-to-date after event handling?)
-        if let Some(txn_id) = local_echo_to_remove {
-            let id = EventItemIdentifier::TransactionId(txn_id.clone());
-            if state.reactions.map.remove(&id).is_none() {
-                warn!(
-                    "Tried to remove reaction by transaction ID, but didn't \
-                     find matching reaction in the reaction map"
-                );
-            }
-        }
-        // Add the remote echo to the reaction_map
-        if let Some(event_id) = remote_echo_to_add {
-            state.reactions.map.insert(
-                EventItemIdentifier::EventId(event_id.clone()),
-                (reaction_sender_data, annotation.clone()),
-            );
-        }
-    }
-
-    state.items.set(idx, timeline_item(new_related, related.internal_id));
-
-    Ok(())
 }
